+++ src/main/java/io/renren/dao/SysUserDao.java
@@ -35,5 +35,5 @@ public interface SysUserDao extends BaseDao<SysUserEntity> {
 	 */
 	int updatePassword(Map<String, Object> map);
 	
-	List <SysUserEntity> queryByParentId(int agencyId);
+	List <SysUserEntity> queryByParentId(Long agencyId);
 }
